requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
drf-yasg>=1.21.6
PyJWT>=2.8.0